pyproject.toml
[project]
name = "dnt"
version = "0.1.0"
description = "Modern Django-Ninja Template with best practices"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Core Django
    "django>=5.2.6",
    "django-ninja>=1.4.3",
    "django-split-settings>=1.3.2",
    # Database
    "psycopg2-binary>=2.9.9",
    "pillow>=10.0.0",
    # Authentication & Security
    "djangorestframework-simplejwt>=5.3.0",
    "django-cors-headers>=4.3.1",
    # Storage
    "django-storages>=1.14.4",
    "boto3>=1.34.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "loguru>=0.7.2",
    # Development
    "django-debug-toolbar>=4.2.0",
    "django-extensions>=3.2.3",
    "ruff>=0.13.2",
    # Caching & Redis
    "redis>=5.0.1",
    "django-redis>=5.4.0",
    # Production server
    "gunicorn>=21.2.0",
    # Testing
    "pytest>=7.4.3",
    "pytest-django>=4.7.0",
    "factory-boy>=3.3.0",
    "pytest-cov>=4.1.0",
    "django-unfold>=0.83.1",
    "uvicorn>=0.42.0",
    "whitenoise>=6.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
]
